Queue for message passing - advantage over shared resources and avoid race condtiion.

+++ thumbnail_maker.py
@@ -4,7 +4,8 @@ import os
 import logging
 from urllib.parse import urlparse
 from urllib.request import urlretrieve
-import threading
+from threading import Thread
+from queue import Queue
 
 import PIL
 from PIL import Image
@@ -17,23 +18,9 @@ class ThumbnailMakerService(object):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming'
         self.output_dir = self.home_dir + os.path.sep + 'outgoing'
-        self.downloaded_bytes = 0
-        self.max_concurrent_downloads = 5
-        # Semaphore is an enhanced version of lock which allows multiple threads to acquire lock at once.
-        # Here, simple example is if Dropbox had a limit on the number of threads which can download the image, we could use a semaphore.
-        # WE have assumed that dropbox has that limit and implemented the semaphore version.
-        self.dl_lock = threading.Semaphore(self.max_concurrent_downloads) 
-
-    def download_image(self, url):
-        # download each image and save to the input dir 
-        with self.dl_lock:
-            logging.info("downloading image at URL " + url)
-            img_filename = urlparse(url).path.split('/')[-1]
-            dest_path = self.input_dir + os.path.sep + img_filename
-            urlretrieve(url, dest_path)
-            img_size = os.path.getsize(dest_path)
-            self.downloaded_bytes += img_size
-            logging.info("image [{} bytes] saved to {}".format(img_size, dest_path))
+        #Instead of having shared global variable between threads which lead to race conditions, let's have a shared queue and use this 
+        # for message passing between the threads.
+        self.img_queue = Queue()
 
     def download_images(self, img_url_list):
         # validate inputs
@@ -44,21 +31,22 @@ class ThumbnailMakerService(object):
         logging.info("beginning image downloads")
 
         start = time.perf_counter()
-        threads = []
         for url in img_url_list:
-            t = threading.Thread(target = self.download_image, args=(url,))
-            t.start()
-            threads.append(t)
+            logging.info("downloading image at URL " + url)
+            img_filename = urlparse(url).path.split('/')[-1]
+            dest_path = self.input_dir + os.path.sep + img_filename
+            urlretrieve(url, dest_path)
+            self.img_queue.put(img_filename)
+        #Poison Pill Technique so that consumer thread knows that end of input has been reached.
+        #Otherwise, as you can see below in the resize method (consumer), the thread will be stuck in infinite loop and won't come out.
+        #Now, consumer can check for a null value in the queue and understand that the EOF has been reached from the producer's side.
+        self.img_queue.put(None) 
         end = time.perf_counter()
 
-        for t in threads:
-            t.join()
-        logging.info("downloaded {} images worth {} bytes in {} seconds".format(len(img_url_list), self.downloaded_bytes, end - start))
+        logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
 
     def perform_resizing(self):
         # validate inputs
-        if not os.listdir(self.input_dir):
-            return
         os.makedirs(self.output_dir, exist_ok=True)
 
         logging.info("beginning image resizing")
@@ -66,22 +54,30 @@ class ThumbnailMakerService(object):
         num_images = len(os.listdir(self.input_dir))
 
         start = time.perf_counter()
-        for filename in os.listdir(self.input_dir):
-            orig_img = Image.open(self.input_dir + os.path.sep + filename)
-            for basewidth in target_sizes:
-                img = orig_img
-                # calculate target height of the resized image to maintain the aspect ratio
-                wpercent = (basewidth / float(img.size[0]))
-                hsize = int((float(img.size[1]) * float(wpercent)))
-                # perform resizing
-                img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
-                
-                # save the resized image to the output dir with a modified file name 
-                new_filename = os.path.splitext(filename)[0] + \
-                    '_' + str(basewidth) + os.path.splitext(filename)[1]
-                img.save(self.output_dir + os.path.sep + new_filename)
+        while True:
+            filename = self.img_queue.get()
+            if filename:
+                logging.info("resizing image {}".format(filename))
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
+                    
+                    # save the resized image to the output dir with a modified file name 
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
 
-            os.remove(self.input_dir + os.path.sep + filename)
+                os.remove(self.input_dir + os.path.sep + filename)
+                self.img_queue.task_done()
+                logging.info("resizing image {} completed".format(filename))
+            else:
+                self.img_queue.task_done()
+                break
         end = time.perf_counter()
 
         logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
@@ -90,9 +86,15 @@ class ThumbnailMakerService(object):
         logging.info("START make_thumbnails")
         start = time.perf_counter()
 
-        self.download_images(img_url_list)
-        # self.perform_resizing()
-
+        t1 = Thread(target=self.download_images, args=([img_url_list]))
+        t2 = Thread(target=self.perform_resizing)
+        t1.start()
+        t2.start()
+        t1.join()
+        t2.join()
+        # Note that here, we are using separate threads for downloading and resiziing, i.e. download and resize operations will be in parallel.
+        # But, still, all the individual download actions would still be sequential. Just that download and resize will be in parallel.
+        # Thus, not much performance gains are expected, only a little bit.
         end = time.perf_counter()
         logging.info("END make_thumbnails in {} seconds".format(end - start))
